requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.3
google-cloud-secret-manager==2.13.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
proto-plus==1.22.3
protobuf==4.21.12
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyspark==3.4.1
pytz==2023.3
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
